Update README.md
Birthday Invitation Page
This GitHub repository hosts a basic HTML web page designed as a birthday invitation. It's an ideal project for beginners diving into the world of web development. The webpage is crafted to be engaging and user-friendly, showcasing essential web development skills through its design and functionality.

Features
Welcoming Message & Event Date: The page greets visitors with a cheerful message and informs them of the event date, immediately setting the tone for a celebration.
Festive Mood with an Image: A carefully chosen image of a birthday cake adds a festive flair to the invitation, visually engaging the guest.
Guest Participation: A list of suggested items for guests to bring encourages participation and adds to the communal spirit of the event.
Venue Information: A convenient Google Maps link makes 

I used fundamental HTML tags such as <h1>, <p>, <ul>, <li>, <img>, and <hr> to structure its content effectively.